Migrate data paths to ~/.local/share/fluorine/, add native build, fix %command% and system Proton scanning
- All data paths migrated from ~/.var/app/com.fluorine.manager/ to
  ~/.local/share/fluorine/ so native and Flatpak builds share the same
  instances, plugins, and configs
- New fluorinepaths.h/.cpp with fluorineDataDir() helper and one-time
  migration from old path (writes MOVED.txt breadcrumb)
- New libs/nak/src/paths.rs as Rust equivalent (data_dir())
- Strip %command% tokens from launch wrapper in protonlauncher.cpp
- Always scan /usr/share/steam/compatibilitytools.d/ for system Proton
  packages (Arch installs Proton there); add Flatpak filesystem permission
- Native build (build-native.sh) installs to ~/.local/share/fluorine/
  with desktop entry and ~/.local/bin symlink instead of portable zip
- build-flatpak.sh wrapper for Flatpak builds
- Fix container locale (LANG=C.UTF-8) for AutoUic warnings
- Fix prefixExists() to handle both compatdata and pfx directory layouts
- Fix globalInstancesRootPath() to use fluorineDataDir() on Linux
- Fix umu-run Flatpak lookup to use fluorineDataDir()
- Update README with build instructions and Arch dependency list

+++ src/src/envmodule.cpp
@@ -529,7 +529,7 @@ HandlePtr Process::openHandleForWait() const
   // On Linux, just wrap the pid. The caller can use waitpid() or
   // kill(pid, 0) to check on the process.
   if (kill(m_pid, 0) == 0) {
-    return HandlePtr(HandleCloser::pointer(m_pid));
+    return HandlePtr(HandleCloser::pointer(static_cast<uintptr_t>(m_pid)));
   }
   return {};
 }
